game_server/backend/requirements.txt
# Flask and Extensions
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-RESTX>=1.3.0
Flask-APScheduler>=1.13.0
Flask-Cors>=4.0.0

# Database
psycopg[binary]>=3.1.0
SQLAlchemy>=2.0.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0
Authlib>=1.3.0

# HTTP Clients (for broadcasting)
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0

